Move import/export/material tests for the model of AliciaSolid. Add VRM model file of AliciaSolid to VRM.Sample/Resources/Models
