+++ Assets/VRM.Samples/Editor/Tests.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 40c1b2cc0da133448a441a01ac4e005b
+folderAsset: yes
+timeCreated: 1546848611
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
