+++ Assets/VRM.Samples/Editor/Tests/VRMImportExportTests.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 81c157c7c98f02742b7af724527af586
+timeCreated: 1532003904
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
