--- Assets/VRM.Samples/Editor/VRMMaterialTests.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 642d2c084bb4b914f9c58fddbf3e13fe
-timeCreated: 1534512276
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
