+++ Assets/VRM.Samples/Resources/Models.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f28857d350eb1b64a807cc2467b75325
+folderAsset: yes
+timeCreated: 1546848586
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
