+++ Assets/VRM.Samples/Resources/Models/.gitignore
@@ -1,0 +1,13 @@
+#
+# This folder contains only VRM files
+#
+*.BlendShapes
+*.BlendShapes.meta
+*.Materials
+*.Materials.meta
+*.Meshes
+*.Meshes.meta
+*.Textures
+*.Textures.meta
+*.prefab
+*.prefab.meta
