+++ Assets/VRM.Samples/Resources/Models/Alicia_vrm-0.40.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 2dd416856c6299848b4b742b3f545ff0
+folderAsset: yes
+timeCreated: 1546848673
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
