+++ Assets/VRM.Samples/Resources/Models/Alicia_vrm-0.40/AliciaSolid_vrm-0.40.vrm.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0973e63dcef234243a5236b99ef74124
+timeCreated: 1546848673
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
